+++ application/controllers/Booking_confirmation.php
@@ -44,8 +44,6 @@ class Booking_confirmation extends EA_Controller
     {
         $appointment_hash = $this->uri->segment(3);
 
-        allow_iframe_embedding(); // Reached from the booking page which may be embedded in an iframe.
-
         $occurrences = $this->appointments_model->get(['hash' => $appointment_hash]);
 
         if (empty($occurrences)) {
@@ -66,6 +64,7 @@ class Booking_confirmation extends EA_Controller
             'matomo_analytics_site_id' => setting('matomo_analytics_site_id'),
             'add_to_google_url' => $add_to_google_url,
             'display_add_to_google_calendar' => setting('display_add_to_google_calendar', '1'),
+            'embedded' => is_embedded_booking_request(),
         ]);
 
         $this->load->view('pages/booking_confirmation');