+++ application/controllers/Localization.php
@@ -32,8 +32,6 @@ class Localization extends EA_Controller
         try {
             method('post');
 
-            allow_iframe_embedding(); // Reached from the booking page which may be embedded in an iframe.
-
             check('language', 'string');
 
             $language = request('language');